pyproject.toml
[tool.poetry]
name = "m11_flask"
version = "0.1.0"
description = ""
authors = ["Yuriy Kuchma <krabatua@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
bcrypt = "^4.0.0"
python-dotenv = {extras = ["cli"], version = "^0.21.0"}
marshmallow = "^3.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
